Feat: order and invite for guest

+++ src/modules/admin/controllers/mediaController.js
@@ -1,11 +1,34 @@
+const Event = require("../models/Event");
+const Media = require("../models/Media");
 const mediaService = require("../services/mediaService");
 
 exports.createMedia = async (req, res) => {
   try {
     const { eventId } = req.params;
+    const event = await Event.findById(eventId);
+    if (!event) return res.status(404).json({ message: "Event not found" });
+
     const mediaData = { ...req.body, eventId };
     const media = await mediaService.createMedia(mediaData);
     res.status(201).json(media);
+  } catch (error) {
+    res.status(400).json({ message: error.message });
+  }
+};
+
+
+exports.createMultipleMedia = async (req, res) => {
+  try {
+    const { eventId } = req.params;
+
+    const event = await Event.findById(eventId);
+    if (!event) {
+      return res.status(404).json({ message: "Event not found" });
+    }
+
+    const mediaData = req.body.map((media) => ({ ...media, eventId }));
+    const media = await mediaService.createMultipleMedia(mediaData);
+    res.status(201).json({ success: true, media });
   } catch (error) {
     res.status(400).json({ message: error.message });
   }
@@ -26,12 +49,22 @@ exports.getAllMedia = async (req, res) => {
   try {
     const { eventId } = req.params;
     const { limit = 10, skip = 0 } = req.query;
-    const media = await mediaService.getAllMedia(
+
+    const parsedLimit = parseInt(limit);
+    const parsedSkip = parseInt(skip);
+
+
+    const { media, totalCount } = await mediaService.getAllMedia(
       eventId,
-      parseInt(limit),
-      parseInt(skip)
+      parsedLimit, parsedSkip
     );
-    res.json(media);
+
+    res.json({
+      media,
+      totalCount,
+      currentPage: Math.floor(parsedSkip / parsedLimit) + 1,
+      totalPages: Math.ceil(totalCount / parsedLimit),
+    });
   } catch (error) {
     res.status(400).json({ message: error.message });
   }
@@ -63,9 +96,23 @@ exports.deleteMultipleMedia = async (req, res) => {
   try {
     const { eventId } = req.params;
     const { mediaIds } = req.body;
+
+    const existingMedia = await Media.find({ _id: { $in: mediaIds }, eventId });
+
+    const existingMediaIds = existingMedia.map((media) => media._id.toString());
+    const missingMediaIds = mediaIds.filter((id) => !existingMediaIds.includes(id));
+
+    if (missingMediaIds.length > 0) {
+      return res.status(404).json({ 
+        message: "Some media items were not found", 
+        missingMediaIds 
+      });
+    }
+
     await mediaService.deleteMultipleMedia(eventId, mediaIds);
-    res.json({ message: "Media items deleted" });
+    res.json({ message: "Media items deleted successfully" });
   } catch (error) {
     res.status(400).json({ message: error.message });
   }
 };
+
